Move only ActionsServer class to make it easier to validate the API
