+++ src/app/clusters/actions-server/app_config_dependent_sources.cmake
@@ -16,8 +16,9 @@
 TARGET_SOURCES(
   ${APP_TARGET}
   PRIVATE
-    "${CLUSTER_DIR}/actions-server.h"
     "${CLUSTER_DIR}/ActionsCluster.cpp"
     "${CLUSTER_DIR}/ActionsDelegate.h"
     "${CLUSTER_DIR}/ActionsStructs.h"
+    "${CLUSTER_DIR}/CodegenIntegration.h"
+    "${CLUSTER_DIR}/actions-server.h"
 )
