+++ src/app/clusters/actions-server/app_config_dependent_sources.gni
@@ -15,5 +15,6 @@ app_config_dependent_sources = [
   "ActionsCluster.cpp",
   "ActionsDelegate.h",
   "ActionsStructs.h",
+  "CodegenIntegration.h",
   "actions-server.h",
 ]
